Placeholders added for the message types Ardupilot is broadcasting
These still need the initial details added to each message type,
prior to a general refactor and clean up.

Also, the message types here are just the ones Ardupilot seems to
be broadcasting in a loop.  I haven't yet attempted to add the
messages it sends in response to a GCS such as QGroundControl.